move test and add whatsnew
